requirements.txt
supabase==2.3.0
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
openai==1.3.0
pandas==2.1.4
numpy==1.24.3
gunicorn==21.2.0

